build v0.02.7
added take damage & death sounds system for player
added new death & take damage sounds for player
fixed rocks model bug
added event frequency manager
fixed camera shake bug
fixed bugs

+++ Assets/Scripts/CommonCore/Damageable/IDamageable.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public interface IDamageable
 {
   
-    public void TakeDamage(int damage, float force, Vector3 hit);
-    public void TakeDamage(int damage);
+    public void TakeDamage(float damage, float force, Vector3 hit);
+    public void TakeDamage(float damage);
     public Transform GetTransform();
 }
