+++ Assets/Scripts/Interactables/Common/ItemDamagable.cs
@@ -15,13 +15,13 @@ public class ItemDamagable : MonoBehaviour,IDamageable
         return transform;
     }
 
-    public void TakeDamage(int damage, float force, Vector3 hit)
+    public void TakeDamage(float damage, float force, Vector3 hit)
     {
             
             OnTakeDamage?.Invoke(force, hit);
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
       
     }
